sideload-repo-systemd: Only remove owned sym links
Check file ownership to ensure flatpak-create-sideload-symlinks.sh
only cleans up links it created. This could be relevant on multi-user
systems with fast user switching.

+++ sideload-repos-systemd/flatpak-create-sideload-symlinks.sh
@@ -26,6 +26,10 @@ done
 
 # Remove any broken symlinks e.g. from drives that were removed
 for f in /run/flatpak/sideload-repos/automount*; do
+    OWNER=$(stat -c '%u' "$f")
+    if [ "$UID" != "$OWNER" ]; then
+        continue
+    fi
     if ! test -e "$f"; then
         rm "$f"
     fi
